kaist-ai-functions/requirements.txt
azure-functions==1.21.3
azure-storage-blob==12.23.1
azure-storage-queue==12.12.0
azure-cosmos==4.7.0
pydantic>=2.0,<3.0
pypdf>=4.0.0
google-genai>=1.0.0
flask>=3.0,<4.0
